Several VS related fixes: - Fixed signed/unsigned warnings appearing when compiling for x86 - Use v110_xp toolset when compiling for x86 to make ETR running on Windows XP
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@336 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ course_render.cpp
@@ -69,7 +69,7 @@ void DrawTrees() {
     TCollidable* treeLocs = &Course.CollArr[0];
     size_t numTrees = Course.CollArr.size();
 
-    for (int i = 0; i< numTrees; i++) {
+    for (size_t i = 0; i< numTrees; i++) {
 		if (clip_course) {
 			if (ctrl->viewpos.z - treeLocs[i].pt.z > fwd_clip_limit) continue;
 		    if (treeLocs[i].pt.z - ctrl->viewpos.z > bwd_clip_limit) continue;
@@ -121,7 +121,7 @@ void DrawTrees() {
 	TItem* itemLocs = &Course.NocollArr[0];
 	size_t numItems = Course.NocollArr.size();
 
-    for (int i = 0; i< numItems; i++) {
+    for (size_t i = 0; i< numItems; i++) {
 		if (itemLocs[i].collectable == 0 || itemLocs[i].drawable == false) continue;
 		if (clip_course) {
 		    if (ctrl->viewpos.z - itemLocs[i].pt.z > fwd_clip_limit) continue;
