+++ gui.h
@@ -140,7 +140,7 @@ void ResetGUI();
 void DrawFrameX (int x, int y, int w, int h, int line,
 			const TColor& backcol, const TColor& framecol, double transp);
 void DrawLevel (int x, int y, int level, double fact);
-void DrawBonus (int x, int y, int max, size_t num);
+void DrawBonus (int x, int y, size_t max, size_t num);
 void DrawBonusExt (int y, size_t numraces, size_t num);
 void DrawCursor ();
